requirements.txt
altair==4.2.0
dash==2.5.1
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
pyrsistent==0.18.1
scipy==1.8.1
streamlit==1.10.0
streamlit_option_menu==0.3.2
